+++ NEWS.md
@@ -1,5 +1,7 @@
 # assemblerr (development version)
 
+
+# assemblerr 0.0.0.9007 
 * added `task` option to `render` function allowing to customize estimation and covariance records, e.g., `render(m, tasks = tsk_estimation(method = "foce"))`
 * added `tsk_ouput()` to specify variables that should be added to $TABLE